pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "thoth-oracle"
version = "0.1.0"
authors = [
    { name="Hobie", email="hobie@example.com" },
]
description = "Thoth Oracle System for XRPL"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "xrpl-py",
    "web3",
    "aiohttp",
    "numpy",
    "pandas",
    "qiskit",
    "pytest",
    "pytest-asyncio",
    "pytest-cov"
]

[project.urls]
"Homepage" = "https://github.com/Hobie1Kenobi/Thoth-Oracle"
"Bug Tracker" = "https://github.com/Hobie1Kenobi/Thoth-Oracle/issues"
